# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple music management API.",
        "title": "Music API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/songs": {
            "get": {
                "description": "Retrieve a list of songs with pagination",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Songs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit the number of songs",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Song"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новую песню в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать новую песню",
                "parameters": [
                    {
                        "description": "Данные для новой песни",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Song"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешный ответ с созданной песней",
                        "schema": {
                            "$ref": "#/definitions/main.Song"
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "put": {
                "description": "Обновляет данные песни по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Обновить информацию о песне",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления песни",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Song"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет песню по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Удалить песню",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/songs/{id}/text": {
            "get": {
                "description": "Возвращает текст песни по ID с поддержкой пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить текст песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID песни",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество куплетов на странице",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorResponse": {
            "description": "Структура для представления ошибок в API",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.NullString": {
            "description": "NullString is a string that can be null",
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid если не пустое",
                    "type": "boolean"
                }
            }
        },
        "main.Song": {
            "description": "Структура для представления информации о песне",
            "type": "object",
            "required": [
                "artist",
                "link",
                "releaseDate",
                "text",
                "title"
            ],
            "properties": {
                "artist": {
                    "description": "Исполнитель песни\n@example \"Artist Name\"",
                    "type": "string"
                },
                "groupName": {
                    "description": "Название группы (может быть null)\n@example \"Group Name\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/main.NullString"
                        }
                    ]
                },
                "id": {
                    "description": "ID песни\n@example 1",
                    "type": "integer"
                },
                "link": {
                    "description": "Ссылка на песню\n@example \"https://example.com/song-link\"",
                    "type": "string"
                },
                "releaseDate": {
                    "description": "Дата выпуска песни\n@example \"2023-11-26\"",
                    "type": "string"
                },
                "text": {
                    "description": "Текст песни\n@example \"This is the text of the song.\"",
                    "type": "string"
                },
                "title": {
                    "description": "Название песни\n@example \"Song Title\"",
                    "type": "string"
                }
            }
        },
        "main.SuccessResponse": {
            "description": "Структура для успешного ответа в API",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}